refactor(helm/v2-alpha): centralize label helpers in labels.go
